+++ commands/reproduce.js
@@ -1,14 +1,11 @@
-const Markov = require("./markov.js")
-var markov = new Markov();
 
 module.exports = {
 	name: "reproduce",
 	description: "Generate a markov generated sentence",
-	help: "Usage: <pre>reproduce",
+	help: "Usage: <pre>reproduce (text)",
 	category: "Markov",
 	execute(message, args) {
-		markov.load("./markovdata/messages.json");
-		message.channel.send(markov.reproduce(Math.floor(Math.random()*200), 4));
+		process.send({command: "reproduce", text: message.content, guild: message.guild.id, channel: message.channel.id});
 	}
 	
 }